refactor llm dd calls
