+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/PlaygroundPage/PlaygroundChat/LLMDropdownOptions.tsx
@@ -16,7 +16,13 @@ import ReactDOM from 'react-dom';
 import {components, OptionProps} from 'react-select';
 
 import {Link} from '../../common/Links';
-import {LLMMaxTokensKey} from '../llmMaxTokens';
+import {TraceObjSchemaForBaseObjectClass} from '../../wfReactInterface/objectClassQuery';
+import {
+  LLM_MAX_TOKENS,
+  LLM_PROVIDER_LABELS,
+  LLMMaxTokensKey,
+} from '../llmMaxTokens';
+import {ProviderStatus} from '../useConfiguredProviders';
 
 export interface ProviderOption {
   label: string | React.ReactNode;
@@ -376,4 +382,89 @@ export const addProviderOption: ProviderOption = {
   ),
   value: 'add-provider',
   llms: [],
+};
+
+export const useLLMDropdownOptions = (
+  configuredProviders: Record<string, ProviderStatus>,
+  configuredProvidersLoading: boolean,
+  customProvidersResult: TraceObjSchemaForBaseObjectClass<'Provider'>[],
+  customProviderModelsResult: TraceObjSchemaForBaseObjectClass<'ProviderModel'>[],
+  customLoading: boolean
+) => {
+  const options: ProviderOption[] = [];
+  const disabledOptions: ProviderOption[] = [];
+
+  if (configuredProvidersLoading) {
+    options.push({
+      label: 'Loading providers...',
+      value: 'loading',
+      llms: [],
+    });
+  } else {
+    Object.entries(configuredProviders).forEach(([provider, {status}]) => {
+      const providerLLMs = Object.entries(LLM_MAX_TOKENS)
+        .filter(([_, config]) => config.provider === provider)
+        .map(([llmKey]) => ({
+          label: llmKey,
+          value: llmKey as LLMMaxTokensKey,
+          max_tokens: LLM_MAX_TOKENS[llmKey as LLMMaxTokensKey].max_tokens,
+        }));
+
+      const option = {
+        label:
+          LLM_PROVIDER_LABELS[provider as keyof typeof LLM_PROVIDER_LABELS],
+        value: provider,
+        llms: status ? providerLLMs : [],
+        isDisabled: !status,
+      };
+
+      if (!status) {
+        disabledOptions.push(option);
+      } else {
+        options.push(option);
+      }
+    });
+  }
+
+  // Add custom providers
+  if (!customLoading) {
+    customProvidersResult?.forEach(provider => {
+      const providerName = provider.val.name || '';
+      const currentProviderModels =
+        customProviderModelsResult?.filter(
+          obj => obj.val.provider === provider.digest
+        ) || [];
+
+      const shortenedProviderLabel =
+        providerName.length > 20
+          ? providerName.slice(0, 2) + '...' + providerName.slice(-4)
+          : providerName;
+
+      const llmOptions = currentProviderModels.map(model => ({
+        label: shortenedProviderLabel + '/' + model.val.name,
+        value: (provider.val.name + '/' + model.val.name) as LLMMaxTokensKey,
+        max_tokens: model.val.max_tokens,
+      }));
+
+      if (llmOptions.length > 0) {
+        options.push({
+          label: providerName,
+          value: providerName,
+          llms: llmOptions,
+        });
+      }
+    });
+  }
+
+  // Combine enabled and disabled options
+  // Add a divider option before the add provider option
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  const allOptions = [
+    ...options,
+    ...disabledOptions,
+    dividerOption,
+    addProviderOption,
+  ];
+
+  return allOptions;
 };
