+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/PlaygroundPage/PlaygroundChat/PlaygroundChatTopBar.tsx
@@ -12,11 +12,14 @@ import {useHistory} from 'react-router-dom';
 import styled from 'styled-components';
 
 import {CopyableId} from '../../common/Id';
+import {TraceObjSchemaForBaseObjectClass} from '../../wfReactInterface/objectClassQuery';
 import {LLMMaxTokensKey} from '../llmMaxTokens';
 import {OptionalTraceCallSchema, PlaygroundState} from '../types';
 import {DEFAULT_SYSTEM_MESSAGE} from '../usePlaygroundState';
 import {LLMDropdown} from './LLMDropdown';
+import {ProviderOption} from './LLMDropdownOptions';
 import {SetPlaygroundStateFieldFunctionType} from './useChatFunctions';
+
 type PlaygroundChatTopBarProps = {
   idx: number;
   settingsTab: number | null;
@@ -27,7 +30,11 @@ type PlaygroundChatTopBarProps = {
   playgroundStates: PlaygroundState[];
   setPlaygroundStates: (playgroundStates: PlaygroundState[]) => void;
   isTeamAdmin: boolean;
-  onConfigureProvider: () => void;
+  refetchConfiguredProviders: () => void;
+  refetchCustomLLMs: () => void;
+  llmDropdownOptions: ProviderOption[];
+  areProvidersLoading: boolean;
+  customProvidersResult: TraceObjSchemaForBaseObjectClass<'Provider'>[];
 };
 
 const DialogActions = styled(MaterialDialogActions)<{$align: string}>`
@@ -47,7 +54,11 @@ export const PlaygroundChatTopBar: React.FC<PlaygroundChatTopBarProps> = ({
   playgroundStates,
   setPlaygroundStates,
   isTeamAdmin,
-  onConfigureProvider,
+  refetchConfiguredProviders,
+  refetchCustomLLMs,
+  llmDropdownOptions,
+  areProvidersLoading,
+  customProvidersResult,
 }) => {
   const history = useHistory();
   const isLastChat = idx === playgroundStates.length - 1;
@@ -130,7 +141,11 @@ export const PlaygroundChatTopBar: React.FC<PlaygroundChatTopBarProps> = ({
           entity={entity}
           project={project}
           isTeamAdmin={isTeamAdmin}
-          onConfigureProvider={onConfigureProvider}
+          refetchConfiguredProviders={refetchConfiguredProviders}
+          refetchCustomLLMs={refetchCustomLLMs}
+          llmDropdownOptions={llmDropdownOptions}
+          areProvidersLoading={areProvidersLoading}
+          customProvidersResult={customProvidersResult}
         />
         {playgroundStates[idx].traceCall?.id && (
           <CopyableId id={playgroundStates[idx]!.traceCall!.id!} type="Call" />
