+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/PlaygroundPage/PlaygroundPage.tsx
@@ -1,12 +1,14 @@
 import {Box} from '@mui/material';
 import {WeaveLoader} from '@wandb/weave/common/components/WeaveLoader';
 import {Pill} from '@wandb/weave/components/Tag/Pill';
-import React, {useEffect, useMemo, useState} from 'react';
+import React, {useCallback, useEffect, useMemo, useState} from 'react';
 
 import {SimplePageLayoutWithHeader} from '../common/SimplePageLayout';
 import {useWFHooks} from '../wfReactInterface/context';
+import {useBaseObjectInstances} from '../wfReactInterface/objectClassQuery';
 import {PlaygroundChat} from './PlaygroundChat/PlaygroundChat';
 import {PlaygroundSettings} from './PlaygroundSettings/PlaygroundSettings';
+import {useConfiguredProviders} from './useConfiguredProviders';
 import {
   DEFAULT_SYSTEM_MESSAGE,
   parseTraceCall,
@@ -82,6 +84,42 @@ export const PlaygroundPageInner = (props: PlaygroundPageProps) => {
     }
   );
 
+  const {
+    result: configuredProviders,
+    loading: configuredProvidersLoading,
+    refetch: refetchConfiguredProviders,
+  } = useConfiguredProviders(props.entity);
+
+  const {
+    result: customProvidersResult,
+    loading: customProvidersLoading,
+    refetch: refetchCustomProviders,
+  } = useBaseObjectInstances('Provider', {
+    project_id: `${props.entity}/${props.project}`,
+    filter: {
+      latest_only: true,
+    },
+  });
+
+  const {
+    result: customProviderModelsResult,
+    loading: customProviderModelsLoading,
+    refetch: refetchCustomProviderModels,
+  } = useBaseObjectInstances('ProviderModel', {
+    project_id: `${props.entity}/${props.project}`,
+    filter: {
+      latest_only: true,
+    },
+  });
+
+  const refetchCustomLLMs = useCallback(() => {
+    refetchCustomProviders();
+    refetchCustomProviderModels();
+  }, [refetchCustomProviders, refetchCustomProviderModels]);
+
+  const areCustomProvidersLoading =
+    customProvidersLoading || customProviderModelsLoading;
+
   useEffect(() => {
     if (!call.loading && call.result) {
       if (call.result.traceCall?.inputs) {
@@ -156,6 +194,13 @@ export const PlaygroundPageInner = (props: PlaygroundPageProps) => {
           setSettingsTab={setSettingsTab}
           settingsTab={settingsTab}
           isOpenInPlayground={!!call.result}
+          configuredProvidersLoading={configuredProvidersLoading}
+          refetchConfiguredProviders={refetchConfiguredProviders}
+          areCustomProvidersLoading={areCustomProvidersLoading}
+          refetchCustomLLMs={refetchCustomLLMs}
+          customProvidersResult={customProvidersResult || []}
+          customProviderModelsResult={customProviderModelsResult || []}
+          configuredProviders={configuredProviders}
         />
       )}
       {settingsTab !== null && (
